requirements.txt
asgiref==3.5.2
dj-database-url==0.5.0
Django==3.0.5
django-heroku==0.3.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
psycopg2==2.9.5
PyJWT==1.7.1
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.6
whitenoise==6.2.0
django-cors-headers==3.5.0
jsonfield==3.1.0
python-dotenv==0.21.0
Pillow==9.3.0
